feat: add zone-aware routing support to memberlist (#762)
**What this PR does**:

In this PR I'm adding zone-aware routing support to memberlist. This
feature relies on `NodeSelectorDelegate` that I've introduced in [this
PR](https://github.com/grafana/memberlist/pull/7).

By default, zone-aware routing is disabled and memberlist client
behaviour does **not** change.

When zone-aware routing is enabled, memberlist client requires the
following information to be set in the config:
- The availability zone
- The node role

There are two roles:
- member: a normal application instance. A member only gossip and
push/pull to other nodes (members and bridges) in **the same zone**.
- bridge: a special application instance that act as a bridge between
AZs. A bridge gossips and push/pull to other nodes in **the same zone**,
plus to **the bridges in other zones**. The bridge is also special
because it prefers to gossip to other bridges in different zones, to
increase the likelihood that messages are actually propagated to the
other zones. The way it works under the hood, is that for each message
to gossip, the bridge always select at least 1 node in other zone (here
comes the preference).

_Draft because I want to integrate it in Mimir and test in a dev cluster
first._

**Which issue(s) this PR fixes**:

Part of https://github.com/grafana/mimir-squad/issues/2634

**Checklist**
- [x] Tests updated

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>