fix(runtimeconfig): fail when can't merge multiple files (#779)
If one file has all string keys in a map, it's a
`map[string]interface{}`, if it has a mix of strings and ints, then it's
`map[interface{}]interface{}`. When two files with different types were
merged, the conflicting keys of the first one were silently overwritten
instead of merging them.

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>