fix: fix TestZoneAwareRouting_EndToEnd flakyness (#785)
**What this PR does**:

This PR fixes `TestZoneAwareRouting_EndToEnd` flakyness. If gossiping is
completely disabled (as in the test "state sync via push/pull") then you
may end up in a split brain situation, unless the periodic rejoin is
enabled (in the test it was not).

In this PR I've changed the order in which nodes are started to
consistently reproduce the split brain scenario, and then I've enabled
rejoin interval to fix it.

Separately, I'm brainstorming if we can end up in a similar situation in
a real world scenario: I don't think so because gossiping is the
primarily method to transfer state (and it doesn't suffer it), but I
will keep thinking about it.

**Which issue(s) this PR fixes**:

Fixes #784 

**Checklist**
- [x] Tests updated

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>