feat(ring/client): add health check grace period to `Pool` (#788)
**What this PR does**:

This PR adds an optional health check grace period to `Pool` in the
`ring/client` package.

The motivating example for this change is in Mimir, where queriers use a
`Pool` to manage connections back to the query-frontend, and monitor
query-frontends' health to remove stale connections from the pool.
Query-frontends are expected to gracefully complete requests on
shutdown, and immediately change their health status as soon as they
receive a shutdown signal (ie. while remaining requests are still
inflight).

However, because query-frontends immediately change their health status
when they receive a shutdown signal, queriers immediately start cleaning
up connections to query-frontends that are shutting down. This means
that any inflight responses from queriers to query-frontends are
aborted, causing queries to fail.

With this grace period in place, queriers can be configured to retain
query-frontend connections until after query-frontends are expected to
have shut down.

Note to reviewers: the tests for this would have been a great candidate
for `testing/synctest`, but we currently target Go 1.24 for this
project, and `synctest` is only considered stable in Go 1.25.

**Which issue(s) this PR fixes**:

(none)

**Checklist**
- [x] Tests updated