fix: configure memberlist bridge nodes to push/pull to 2 nodes per cycle (#798)
**What this PR does**:

The bridge always prefer another bridge as first node. If the bridge
only push/pull to 1 node per interval, then it will only communicate to
bridges, potentially leading to network partitioning if the gossiping is
not working to propagate changes.

To reduce the likelihood of network partitioning when gossiping is not
working and periodic push/pull is enabled, in this PR I configure the
bridge to push/pull to 2 nodes per interval (the first node is a bridge,
and the second node is selected randomly).

This change removes the need of enabling periodic rejoin from
`TestZoneAwareRouting_EndToEnd` because now we don't end up with network
partitioning when gossiping is disabled, and the state is propagated
exclusively using push/pull.

~~This PR is based on https://github.com/grafana/memberlist/pull/11.
`go.mod` will be updated after memberlist PR will be merged.~~

**Which issue(s) this PR fixes**:

N/A

**Checklist**
- [x] Tests updated

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>