Use a more clever zone lookup
slices.Contains is the slowest part of the new SelectNodes(), so I
thought: maybe we can do better?

Realistically, most of the setups will have zone-[abc], or
similar, and after checking a couple of nodes we'll already have all
three zones in the slice, so if we managed to keep the slice sorted, we
can check the expected index in the slice first.

This does a 15% decrease in cpu time for
BenchmarkZoneAwareNodeSelectionDelegate_SelectNodes

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>