feat(ring): add ring_zone_member metric (#797)
**What this PR does**:
Adds `ring_zone_members` metric, that is exposed when the ring runs in
zone-aware mode. On startup, the metric is empty. If instances from a
specific zone join the ring, the list of `allRingZones` will be
populated. If the last and only instance of a zone leaves
`ring_zone_members` is set to `0`.

**Checklist**
- [x] Tests updated